+++ apps/backend/src/main/kotlin/hu/kirdev/foodex/dto/ApplyForShiftDTO.kt
@@ -2,6 +2,6 @@ package hu.kirdev.foodex.dto
 import jakarta.validation.constraints.NotBlank
 
 data class ApplyForShiftDTO(
-    @field:NotBlank val userId: Long,
-    @field:NotBlank val shiftId: Long
+    @field:NotBlank val userId: Int,
+    @field:NotBlank val shiftId: Int
 )
