+++ apps/backend/src/main/kotlin/hu/kirdev/foodex/repository/ConfigurationRepository.kt
@@ -3,6 +3,6 @@ package hu.kirdev.foodex.repository
 import hu.kirdev.foodex.model.ConfigurationEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ConfigurationRepository : JpaRepository<ConfigurationEntity, Long> {
+interface ConfigurationRepository : JpaRepository<ConfigurationEntity, Int> {
     fun findTopByOrderByIdDesc(): ConfigurationEntity?
 }
